chore(): rename task and test files
