+++ src/engine/create-changeset/tasks/index.ts
@@ -1,8 +1,8 @@
-import { affectedContentTypesDivergedTask } from './create-affected-content-types-diverged-task'
-import { computeIdsTask } from './create-compute-ids-task'
-import { fetchPartialEntitiesTask } from './create-entities-task'
-import { fetchAddedEntitiesTask } from './create-fetch-added-entities-task'
-import { fetchChangedTasks } from './create-fetch-changed-tasks'
+import { affectedContentTypesDivergedTask } from './affected-content-types-diverged-task'
+import { computeIdsTask } from './compute-ids-task'
+import { fetchPartialEntitiesTask } from './fetch-partial-entities-task'
+import { fetchAddedEntitiesTask } from './fetch-added-entities-task'
+import { fetchChangedTasks } from './fetch-changed-tasks'
 
 export const createChangesetTasks = {
   fetchPartialEntitiesTask,
